requirements.txt
langchain==0.2.5
langchain-community==0.2.5
langchain-chroma==0.1.1
langchain-huggingface==0.0.3
langchainhub
ipykernel
python-dotenv
pypdf
beautifulsoup4
arxiv
pymupdf
wikipedia
langchain-openai==0.1.8
sentence-transformers
faiss-cpu
chromadb
streamlit
langchain-groq==0.1.6
langchain-core==0.2.10
fastapi
langserve
ddgs
duckdb
pandas
mysql-connector-python
SQLAlchemy==2.0.30
validators==0.28.1
youtube-transcript-api
unstructured
pytube
pydantic==2.7.4
numexpr
lxml
gradio
PyPDF2